requirements.txt
opencv-python-headless
requests
python-dotenv
google-generativeai
google-genai